Audio menu in
Sound on/off works, levels can be changed by 1-5 number keys, going to work on using arrow keys as well as mouse selection as well but have to head off reffing. Volume switching DOES NOT WORK at the moment. Need to caputre the channel and update the volume there but don't know how at this time